Extract Sink to its own module
